pyproject.toml
[project]
name = "snowpolars"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "adbc-driver-snowflake>=1.4.0",
    "polars>=1.22.0",
    "pyarrow>=19.0.0",
    "python-dotenv>=1.0.1",
]
